scale: Switch to kernel driver for sampling hx711
User space bit banging python driver is not stable.  It worked OK
on a Raspberry Pi 3, but once I switched to a Pi Zero, it was mostly
garbage data.  The bit banged clock line can't meet the tight timing
requirement of the HX711.  Switching to the kernel driver for this
gives much more reliable data on lightweight systems like Pi Zero.