arduino: Create arduino BLE scale reader
After multiple failures trying to reliably read HX711 data using
a Raspberry Pi, I gave up and switched to an Arduino.  The intermittant
latency on the RPi was causing corruption on the HX711 readings.

I've switched to using an Adafruit Feature 32u4 Bluefruit LE to
interface with the HX711.  It periodicly takes readings.  When it
detects a weight change, or if enough time has elapsed, then it
broadcasts the reading using the WeightScale GATT profile in BLE.

The RPi can connect via BLE now and reliable get the readings and
pass them up to AWS.